added the prefix of the next month. and moved the code to its own js file.
