+++ src/assets/js/months.js
@@ -108,3 +108,57 @@ export const monthList = [
 
         return updatedMonthList;
     }
+
+
+export const getDaysFromPreviousMonth = (selectedYear, months, selectedMonth) => {
+    //get the prefix of the last month
+
+    const todayDate = new Date(selectedYear, months.indexOf(selectedMonth), 1);
+    todayDate.setDate(0); // Set the date to the last day of the previous month
+    const daysInPreviousMonth = todayDate.getDate(); // Get the number of days in the previous month
+    const prefixDaysArray = [];
+
+    for (
+      let i = daysInPreviousMonth;
+      i > daysInPreviousMonth - todayDate.getDay();
+      i--
+    ) {
+      prefixDaysArray.unshift(i);
+    }
+
+    return prefixDaysArray;
+  };
+
+
+export const getNextMonthDays = (selectedYear, months, selectedMonth) => {
+    //get the prefix of the next month
+
+    const nextMonthDate = new Date( //next month date
+      selectedYear,
+      months.indexOf(selectedMonth) + 1,
+      1
+    );
+    
+    const todayDate = new Date( //todays date
+        selectedYear,
+        months.indexOf(selectedMonth),
+        1
+      );
+      todayDate.setDate(0); //setting to the last day of the previous month.
+
+    const currentMonthLastDay = new Date( //current last day of this month.
+      selectedYear,
+      months.indexOf(selectedMonth) + 1,
+      1
+    );
+    currentMonthLastDay.setDate(0); // Set the date to the last day of the previous month
+    const nextMonthPrefixArray = [];
+    for (
+      let i = nextMonthDate.getDate();
+      i <= 42 - (currentMonthLastDay.getDate() + todayDate.getDay()); //the display grid is 42 grids in total and minus the maximum date of the current month plus the last day of the previous month (Tues(2))
+      i++
+    ) {
+      nextMonthPrefixArray.push(i);
+    }
+    return nextMonthPrefixArray;
+  };
